use %lu instead of casting to uint

+++ gtkgame.c
@@ -5566,7 +5566,7 @@ extern void SetRollouts( gpointer p, guint n, GtkWidget *pwIgnore )
 
     if( abs(rw.rcRollout.nSeed) != abs(rcRollout.nSeed) ) {
 	    /* seed may be unsigned long int */
-      sprintf( sz, "set rollout seed %u", (unsigned int)rw.rcRollout.nSeed );
+      sprintf( sz, "set rollout seed %lu", rw.rcRollout.nSeed );
       UserCommand( sz );
     }
